src/evolite/core/evolution.py
from typing import List, Callable, Optional
import numpy as np
from ..models.neural_net import CompressibleNetwork
from ..utils.helpers import evaluate_fitness


class EvolutionaryOptimizer:
    """Evolutionary algorithm for neural network compression."""

    def __init__(
        self,
        population_size: int = 50,
        mutation_rate: float = 0.1,
        crossover_rate: float = 0.8,
        selection_pressure: float = 0.2,
    ):
        """
        Initialize the evolutionary optimizer.

        Args:
            population_size: Number of individuals in population
            mutation_rate: Probability of mutation per gene
            crossover_rate: Probability of crossover
            selection_pressure: Selection pressure for tournament
        """
        self.population_size = population_size
        self.mutation_rate = mutation_rate
        self.crossover_rate = crossover_rate
        self.selection_pressure = selection_pressure
        self.current_generation = 0
        self.population: List[CompressibleNetwork] = []

    def evolve(
        self,
        network: CompressibleNetwork,
        fitness_fn: Callable,
        n_generations: int = 100,
    ) -> CompressibleNetwork:
        """
        Evolve the network to find optimal compression.

        Args:
            network: Initial network to optimize
            fitness_fn: Function to evaluate individual fitness
            n_generations: Number of generations to evolve

        Returns:
            CompressibleNetwork: Best compressed network
        """
        raise NotImplementedError("Coming soon!")
